wip repacking: methods implemented
